+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.0.3 - 2021-12-04
 
-- Fixed Team Burning Systems Omega Belt potentially being loaded twice
+- Fixed Team Burning Systems' Omega Belt potentially being loaded twice
 
 ## v1.0.2 - 2021-07-22