+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- Build map list dynamically
+
 ## v1.0.3 - 2021-12-04
 
 - Fixed Team Burning Systems' Omega Belt potentially being loaded twice